# legalai-backend/app/static/openapi.yaml
openapi: 3.0.3
info:
  title: Legal Lawyer AI API
  version: "1.0.0"
  description: |
    API specification for Legal Lawyer AI backend.

    Safe Mode:
    - Only endpoints that declare the `X-Safe-Mode` header enforce it.
    - When `X-Safe-Mode: 1`, those endpoints block writes and return 403.
servers:
  - url: https://gowaned-beckham-unlawful.ngrok-free.dev
  - url: http://127.0.0.1:5000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    safeModeHeader:
      type: apiKey
      in: header
      name: X-Safe-Mode
      description: "Set to '1' to enable safe mode (read-only, no persistence)"
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
    Error:
      type: object
      properties:
        message: { type: string }
        error: { type: string }
        details: { $ref: "#/components/schemas/AnyObject" }
...
    NotificationItem:
      type: object
      required: [id, type, title, body, data, scope, createdAt, isRead]
      properties:
        id: { type: integer, format: int64 }
        type: { type: string, example: "content_update" }
        title: { type: string }
        body: { type: string }
        data: { type: object, additionalProperties: true }
        scope: { type: string, enum: [broadcast, user] }
        createdAt: { type: string, format: date-time, nullable: true }
        isRead: { type: boolean }
        readAt: { type: string, format: date-time, nullable: true }
    NotificationsListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/NotificationItem" }
    NotificationPreferences:
      type: object
      required: [contentUpdates, lawyerUpdates, reminderNotifications]
      properties:
        contentUpdates: { type: boolean }
        lawyerUpdates: { type: boolean }
        reminderNotifications: { type: boolean }
    NotificationPreferencesUpdate:
      type: object
      properties:
        contentUpdates: { type: boolean }
        lawyerUpdates: { type: boolean }
        reminderNotifications: { type: boolean }
    NotificationDeviceTokenRequest:
      type: object
      required: [platform, token]
      properties:
        platform:
          type: string
          enum: [android, ios]
        token:
          type: string
          minLength: 10
    NotificationUnregisterRequest:
      type: object
      required: [token]
      properties:
        token: { type: string, minLength: 10 }

    # Error Schemas (Industry Best Practice)
...
    UnauthorizedErrorResponse:
      type: object
      required: [message]
      properties:
        message: { type: string, example: "Invalid token" }
        error: { type: string, example: "Unauthorized" }
    ForbiddenErrorResponse:
      type: object
      required: [message]
      properties:
        message: { type: string, example: "Admin only" }
        error: { type: string, example: "Forbidden" }
...
security: [] # default none; per-route add as needed
paths:
...
  /api/v1/notifications:
    get:
      tags: [Notifications]
      summary: List notifications
      description: Returns the latest notifications (newest first).
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: before
          schema: { type: integer }
          description: "Return items with id < before (pagination)"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationsListResponse" }
        "400":
          description: Bad Request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
  /api/v1/notifications/{notificationId}/read:
    post:
      tags: [Notifications]
      summary: Mark notification as read
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: notificationId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "400":
          description: Bad Request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
  /api/v1/notifications/mark-all-read:
    post:
      tags: [Notifications]
      summary: Mark all notifications as read
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [ok, updated]
                properties:
                  ok: { type: boolean, example: true }
                  updated: { type: integer, example: 5 }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
  /api/v1/notifications/unread-count:
    get:
      tags: [Notifications]
      summary: Get unread notifications count
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count: { type: integer, example: 2 }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
  /api/v1/notifications/preferences:
    get:
      tags: [Notifications]
      summary: Get notification preferences
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPreferences" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
    put:
      tags: [Notifications]
      summary: Update notification preferences
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NotificationPreferencesUpdate" }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationPreferences" }
        "400":
          description: Bad Request
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
  /api/v1/notifications/register-device-token:
    post:
      tags: [Notifications]
      summary: Register device token
      description: Registers FCM/APNS token for push notifications.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NotificationDeviceTokenRequest" }
      responses:
        "200":
          description: Token registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "400":
          description: Invalid platform or missing token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
  /api/v1/notifications/unregister-device-token:
    post:
      tags: [Notifications]
      summary: Unregister device token
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/NotificationUnregisterRequest" }
      responses:
        "200":
          description: Token unregistered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
        "400":
          description: Missing token
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }

  # -------------------- ADMIN KNOWLEDGE --------------------
...
  /api/v1/admin/notifications:
    get:
      tags: [Admin]
      summary: List admin notifications (Admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: limit
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
        - in: query
          name: before
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationsListResponse" }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForbiddenErrorResponse" }
  /api/v1/admin/notifications/{notificationId}/read:
    post:
      tags: [Admin]
      summary: Mark admin notification as read (Admin only)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: notificationId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean, example: true }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForbiddenErrorResponse" }
        "404":
          description: Not Found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/admin/notifications/unread-count:
    get:
      tags: [Admin]
      summary: Get unread admin notifications count (Admin only)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count: { type: integer, example: 3 }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UnauthorizedErrorResponse" }
        "403":
          description: Forbidden
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ForbiddenErrorResponse" }
...